OTP code working (through console for temporary)

+++ utils/otp.php
@@ -29,11 +29,48 @@ class OTPError {
     }
 
     public function otp($value): bool {
-        $userData = getOtpUser("otp_number", $value);
+        $value = trim($value);
 
+        if (!$value) {
+            array_push($this->errorMessages, "Kode verifikasi wajib diisi");
+            return false;
+        }
+
+        $userData = getOtpUser("otp_number", $value);
         if (!$userData) {
             array_push($this->errorMessages, "Kode verifikasi salah");
             return false;
+        }
+        return true;
+    }
+
+    public function password($password, $cpassword): bool {
+        $password = trim($password);
+        $cpassword = trim($cpassword);
+
+        if (!($password && $cpassword)) {
+            array_push($this->errorMessages, "Semua kolom wajib diisi");
+            return false;
+        }
+        if ($password !== $cpassword) {
+            array_push($this->errorMessages, "Konfirmasi password harus sama dengan password");
+            return false;
+        }
+        if (!(strlen($password) >= 8 && strlen($cpassword) >= 8)) {
+            array_push($this->errorMessages, "Password minimal 8 karakter");
+            return false;
+        }
+        return true;
+    }
+
+    public function validateUser($email, $otp): bool {
+        $user = getOtpUser("email", $email);
+        if (!$user) {
+            array_push($this->errorMessages, "Email salah");
+            return false;
+        }
+        if ($user["otp_number"] !== "$otp") {
+            array_push($this->errorMessages, "Kode verifikasi salah");
         }
         return true;
     }
@@ -49,11 +86,29 @@ function getOtpEmailErrors($email) {
     return $error->getErrorMessages();
 }
 
-function getOtpUser($column, $columnValue) {
+function getOtpNumberErrors($otp) {
+    $error = new OTPError();
+    $error->otp($otp);
+    return $error->getErrorMessages();
+}
+
+function getOTPUserErrors($email, $otp) {
+    $error = new OTPError();
+    $error->validateUser($email, $otp);
+    return $error->getErrorMessages();
+}
+
+function getOtpPasswordErrors($password, $cpassword) {
+    $error = new OTPError();
+    $error->password($password, $cpassword);
+    return $error->getErrorMessages();
+}
+
+function getOtpUser($column, $columnValue, $all=false) {
     global $connection;
     $sql = "SELECT * FROM otp WHERE $column LIKE '$columnValue'";
     $query = $connection->query($sql);
-    $result = $query->fetch_assoc();
+    $result = $all ? $query->fetch_all(MYSQLI_ASSOC) : $query->fetch_assoc();
     return $result;
 }
 
@@ -88,6 +143,11 @@ function deleteExpiredOtpUser($dateNow) {
     $dateNow = $dateNow;
     $selectedQuery = $connection->query("SELECT expire_date FROM otp");
     $selected = $selectedQuery->fetch_all(MYSQLI_ASSOC);
+
+    if (!count($selected)) {
+        return json_encode(["error" => false]);
+    }
+
     $filteredExpires = array_filter($selected, function($var) {
         global $dateNow;
         return intval($var["expire_date"]) < intval($dateNow);
@@ -96,6 +156,11 @@ function deleteExpiredOtpUser($dateNow) {
         return "'" . $var["expire_date"] . "'";
     }, $filteredExpires);
     $implodedExpireDates = implode(", ", $expireDates);
+
+    if (!$implodedExpireDates) {
+        return json_encode(["error" => false]);
+    }
+    
     $sql = "DELETE FROM otp WHERE expire_date IN ($implodedExpireDates)";
     $connection->query($sql);
     return json_encode(["error" => false]);
